fix update-account
