+++ src/utils/Var.tsx
@@ -1,1 +1,1 @@
-export const MODULE_ADDRESS="c77e80ec8b348f203615f6cf711567c87f5b6ea19e1c40799fb6480d092f02fd"
+export const MODULE_ADDRESS="fd874114ae9adacd5ee56e5ccdfd626b15bc1cabd18b24d2369a1664ec60244e"
